fix(clippy): resolve all clippy warnings in cpd-finder and cpd-reporter
- blame.rs: use strip_prefix instead of starts_with+manual slice (manual_strip)
- blame.rs: change &mut Vec<CpdClone> to &mut [CpdClone] (ptr_arg)
- orchestrate.rs: use MULTI_FORMAT_EXTS.contains() (manual_contains)
- orchestrate.rs: use .flatten() instead of .filter_map(|x| x) (filter_map_identity)
- statistics.rs: remove redundant millis as u32 cast (unnecessary_cast)
- walker.rs: derive Default instead of manual impl (derivable_impls)
- xml_reporter.rs: remove needless borrow in BytesCData::new() (needless_borrows_for_generic_args)

+++ rust/crates/cpd-finder/src/statistics.rs
@@ -85,7 +85,7 @@ use std::time::{SystemTime, UNIX_EPOCH};
         let duration = SystemTime::now().duration_since(UNIX_EPOCH).unwrap_or_default();
         let secs = duration.as_secs();
         let millis = duration.subsec_millis();
-        chrono::DateTime::from_timestamp(secs as i64, (millis as u32) * 1_000_000)
+        chrono::DateTime::from_timestamp(secs as i64, millis * 1_000_000)
             .map(|dt| dt.format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string())
             .unwrap_or_else(|| format!("{secs}"))
     };
